feat: configure project for deployment

+++ src/utils/system_utils.py
@@ -8,8 +8,8 @@ def create_path(path):
         os.makedirs(path)
 
 def clean_up(input_file, output_file):
-        if os.path.exists(output_file):
+        if output_file and os.path.exists(output_file):
             os.remove(output_file)
     
-        if os.path.exists(input_file) and not os.path.exists(output_file):
+        if input_file and os.path.exists(input_file):
             os.remove(input_file)
